feat: [cli] add hello world functionality
Create `console.py` file adding the hello world feature using click. Additionally, pyproject.toml was changed by creating the cli script to be ran using `poetry run cli`. To make it work, we needed to change the name of the overall package (from templyte to cli) because there was no subpackage named `templayte`. Since we created the cli folder, it was a good choice to name it that way.